chore: stop enriching window
Adding properties to global window object is considered bad practice.
This lib added property to  get an accessible reference to the event queue
at all time.
An event-queue is now instantiated, setup, and exported by the EventQueue module.
An import is enough to use it everywhere.

Fix #74
